# [Model Monitoring] Edit controller for better performance