--- a/tests/system/model_monitoring/test_app.py
+++ b/tests/system/model_monitoring/test_app.py
@@ -330,7 +330,7 @@ class _V3IORecordsChecker:
 class TestMonitoringAppFlow(TestMLRunSystem, _V3IORecordsChecker):
     project_name = "test-app-flow-v4"
     # Set image to "<repo>/mlrun:<tag>" for local testing
-    image: typing.Optional[str] = "quay.io/davesh0812/mlrun:1.8.2"
+    image: typing.Optional[str] = None
     error_count = 10
 
     @classmethod
@@ -691,7 +691,7 @@ class TestMonitoringAppFlow(TestMLRunSystem, _V3IORecordsChecker):
             future = executor.submit(self._deploy_model_serving, with_training_set)
 
         serving_fn = future.result()
-        # self._add_error_alert()
+        self._add_error_alert()
 
         time.sleep(5)
         last_request = self._infer(
@@ -726,7 +726,7 @@ class TestMonitoringAppFlow(TestMLRunSystem, _V3IORecordsChecker):
         self._test_api(ep_id=mep.metadata.uid)
         if _DefaultDataDriftAppData in self.apps_data:
             self._test_model_endpoint_stats(mep=mep)
-        # self._test_error_alert()
+        self._test_error_alert()
 
 
 @TestMLRunSystem.skip_test_if_env_not_configured