Few issue I address in this pr : 

1. Initial the `MonitoringApplicationController` in the pod initialization and not in each invoke, educing execution time by 300 milliseconds per run.
2. Send regular event only if the `last_request` is greater when `last_analyze`, to reduce non relevant traffic.
3. Send NOP event according to the minimal `last_analyze`.
4. Add thread pool to the chief logic.

[ML-8111](https://iguazio.atlassian.net/browse/ML-8111)


[ML-8111]: https://iguazio.atlassian.net/browse/ML-8111?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

```
With 100 model endpoints

No app trigger
> 2025-01-26 14:01:22,260 [info] Starting monitoring controller chief
> 2025-01-26 14:01:26,933 [info] Finishing monitoring controller chief
> 2025-01-26 14:01:28,469 [info] Finish analyze for: {"timestamp":"2025-01-26 13:51:46.208707+00:00"} (last worker finish)

= 6sec


With app trigger( include nops):

> 2025-01-26 14:11:23,095 [info] Starting monitoring controller chief
> 2025-01-26 14:11:27,989 [info] Finishing monitoring controller chief
> 2025-01-26 14:11:38,314 [info] Finish analyze for: {"timestamp":"2025-01-26T14:11:33.164071+00:00"}(last worker finish)

=15 sec


With 1000 model endpoints:

No app trigger
> 2025-01-26 14:33:45,192 [info] Starting monitoring controller chief
> 2025-01-26 14:33:58,578 [info] Finishing monitoring controller chief
> 2025-01-26 14:34:00,658 [info] Finish analyze for: {"timestamp":"2025-01-26 14:27:56.710203+00:00"}(last worker finish)

= 15 sec

With app trigger:
> 2025-01-26 14:43:45,095 [info] Starting monitoring controller chief
> 2025-01-26 14:44:07,942 [info] Finishing monitoring controller chief
> 2025-01-26 14:44:15,124 [info] Finish analyze for: {"timestamp":"2025-01-26T14:44:15.092389+00:00"}(last worker finish)

= 30 sec
```